Use containerd content store in SOCI CLI
This chnage uses the containerd content store for the SOCI CLI. SOCI
indexes will be tied to the image lifecycle and will get automtaically
garbage collected when the image is removed.

The snapshotter still uses the local soci content store by default until
we're confident that the circular reference back to contaienrd is safe.

Signed-off-by: Kern Walster <walster@amazon.com>